Update sammyPymote.py
Used the network helper class instead of using hard coded variables about local interface ip and mac address. This way the library gets more dynamic. The network library makes the assumption that the primary interface of the pc that connects to the internet is also the one connecting to the tv. If that assumption is wrong, well, fix the code, or use hard coded variables.